requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
psutil==5.9.6
speedtest-cli==2.1.3
requests==2.31.0
ping3==4.0.8